Update requirements.txt
Hi, I have no idea if this is either a rude or polite thing to do, but I noticed a small missing dependency (for the jupyter notebook).
In any case, can't wait to play with this. Thank you for the repo!
(again, sorry if this is a rude thing to do. I've never tried to suggest a change to a stranger before, and have no idea what the conventions are.)